Refactor: 모달 -> 포커스로 이름 변경
